// src/lib/api/model/leadResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { LeadResponseDtoSocialLinks } from "./leadResponseDtoSocialLinks";
import type { LeadResponseDtoStatus } from "./leadResponseDtoStatus";
import type { LeadResponseDtoPriority } from "./leadResponseDtoPriority";
import type { LeadResponseDtoCurrency } from "./leadResponseDtoCurrency";
import type { LeadResponseDtoConversation } from "./leadResponseDtoConversation";

export interface LeadResponseDto {
  id: string;
  organizationId: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  country: string;
  timezone: string;
  primaryLanguage: string;
  preferredLanguage?: string;
  nativeName?: string;
  email?: string;
  socialLinks?: LeadResponseDtoSocialLinks;
  gender?: string;
  sourceId?: string;
  pipelineStageId?: string;
  assignedAgentId?: string;
  status: LeadResponseDtoStatus;
  priority: LeadResponseDtoPriority;
  estimatedValue?: number;
  currency?: LeadResponseDtoCurrency;
  expectedServiceDate?: string;
  nextFollowUpAt?: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
  conversation?: LeadResponseDtoConversation;
}